Quick backup

+++ scripts/drivers/wrf.sh
@@ -104,21 +104,21 @@ fi
 # Options below are defined in workflow variables 
 #
 # MEMID        = Ensemble ID index, 00 for control, i > 00 for perturbation
-# BKG_DATA     = String case variable for supported inputs: GFS, GEFS currently
-# BKG_INT      = Interval of input data in HH
 # STRT_TIME    = Simulation start time in YYMMDDHH
 # IF_DYN_LEN   = "Yes" or "No" switch to compute forecast length dynamically 
 # FCST_HRS     = Total length of WRF forecast simulation in HH, IF_DYN_LEN=No
 # EXP_VRF      = Verfication time for calculating forecast hours, IF_DYN_LEN=Yes
-# WRFOUT_INT   = Interval of wrfout in HH
-# CYCLE_INT    = Interval in HH on which DA is cycled in a cycling control flow
-# CYCLE_TIME   = Simulation cycle date time in YYMMDDHH
+# BKG_INT      = Interval of input data in HH
+# BKG_DATA     = String case variable for supported inputs: GFS, GEFS currently
 # MAX_DOM      = Max number of domains to use in namelist settings
 # DOWN_DOM     = First domain index to downscale ICs from d01, set parameter
 #                less than MAX_DOM if downscaling to be used
-# IF_CYCLING   = Yes / No: whether to use ICs / BCs from GSI / WRFDA analysis
-#                or real.exe, case insensitive
-# IF_RESTART   = 
+# WRFOUT_INT   = Interval of wrfout in HH
+# CYCLE_INT    = Interval in HH on which DA is cycled in a cycling control flow
+# WRF_IC       = Defines where to source WRF initial and boundary conditions from
+#                  WRF_IC = REALEXE : ICs / BCs from CYCLE_HME/realprd
+#                  WRF_IC = CYCLING : ICs / BCs from GSI / WRFDA analysis
+#                  WRF_IC = RESTART : ICs from restart file in CYCLE_HME/wrfprd
 # IF_SST_UPDTE = Yes / No: whether WRF uses dynamic SST values 
 # IF_FEEBACK   = Yes / No: whether WRF domains use 1- or 2-way nesting
 #
@@ -132,23 +132,8 @@ else
   memid=`printf %02d $(( 10#${MEMID} ))`
 fi
 
-if [[ ${BKG_DATA} != GFS &&  ${BKG_DATA} != GEFS ]]; then
-  msg="ERROR: \${BKG_DATA} must equal 'GFS' or 'GEFS'"
-  msg+=" as currently supported inputs."
-  echo ${msg}
-  exit 1
-fi
-
-if [ ! ${BKG_INT} ]; then
-  echo "ERROR: \${BKG_INT} is not defined."
-  exit 1
-elif [ ! ${BKG_INT} -gt 0 ]; then
-  echo "ERROR: \${BKG_INT} must be HH > 0 for the frequency of data inputs."
-  exit 1
-fi
-
 if [ ${#STRT_TIME} -ne 10 ]; then
-  echo "ERROR: \${STRT_TIME}, '${STRT_TIME}', is not in 'YYYYMMDDHH' format." 
+  echo "ERROR: \${STRT_TIME}, ${STRT_TIME}, is not in 'YYYYMMDDHH' format." 
   exit 1
 else
   # Convert STRT_TIME from 'YYYYMMDDHH' format to strt_time Unix date format
@@ -168,7 +153,7 @@ if [[ ${IF_DYN_LEN} = ${NO} ]]; then
 elif [[ ${IF_DYN_LEN} = ${YES} ]]; then
   echo "Generating forecast forcing data until experiment validation time."
   if [ ${#EXP_VRF} -ne 10 ]; then
-    echo "ERROR: \${EXP_VRF}, `${EXP_VRF}` is not in 'YYYMMDDHH' format."
+    echo "ERROR: \${EXP_VRF}, ${EXP_VRF} is not in 'YYYMMDDHH' format."
     exit 1
   else
     # compute forecast length relative to start time and verification time
@@ -185,40 +170,19 @@ fi
 # define the end time based on forecast length control flow above
 end_time=`date -d "${strt_time} ${fcst_len} hours"`
 
-if [ ${#WRFOUT_INT} -ne 2 ]; then
-  echo "ERROR: \${WRFOUT_INT} is not in HH format."
+if [ ! ${BKG_INT} ]; then
+  echo "ERROR: \${BKG_INT} is not defined."
   exit 1
-elif [ ! ${WRFOUT_INT} -gt 0 ]; then
-  echo "ERROR: \${WRFOUT_INT} must be an integer for the max WRF domain index > 0." 
-  exit 1
-fi
-
-if [ ! "${CYCLE_INT}" ]; then
-  echo "ERROR: \${CYCLE_INT} is not defined."
+elif [ ! ${BKG_INT} -gt 0 ]; then
+  echo "ERROR: \${BKG_INT} must be HH > 0 for the frequency of data inputs."
   exit 1
 fi
 
-if [ ! "${STRT_TIME}" ]; then
-  echo "ERROR: \${STRT_TIME} is not defined."
+if [[ ${BKG_DATA} != GFS &&  ${BKG_DATA} != GEFS ]]; then
+  msg="ERROR: \${BKG_DATA} must equal 'GFS' or 'GEFS'"
+  msg+=" as currently supported inputs."
+  echo ${msg}
   exit 1
-fi
-
-# Convert STRT_TIME from 'YYYYMMDDHH' format to strt_time Unix date format
-if [ ${#STRT_TIME} -ne 10 ]; then
-  echo "ERROR: \${STRT_TIME}, '${STRT_TIME}', is not in 'YYYYMMDDHH' format." 
-  exit 1
-else
-  strt_time="${STRT_TIME:0:8} ${STRT_TIME:8:2}"
-fi
-strt_time=`date -d "${strt_time}"`
-end_time=`date -d "${strt_time} ${fcst_len} hours"`
-
-# Convert CYCLE_TIME from 'YYYYMMDDHH' format to strt_time Unix date format
-if [ ${#CYCLE_TIME} -ne 10 ]; then
-  echo "ERROR: cycle time, '${CYCLE_TIME}', is not in 'YYYYMMDDHH' format."
-  exit 1
-else
-  cycle_time="${CYCLE_TIME:0:8} ${CYCLE_TIME:8:2}"
 fi
 
 if [ ! "${MAX_DOM}" ]; then
@@ -231,8 +195,31 @@ if [ ! "${DOWN_DOM}" ]; then
   exit 1
 fi
 
-if [[ ${IF_CYCLING} != ${YES} && ${IF_CYCLING} != ${NO} ]]; then
-  echo "ERROR: \${IF_CYCLING} must equal 'Yes' or 'No' (case insensitive)."
+if [ ${#WRFOUT_INT} -ne 2 ]; then
+  echo "ERROR: \${WRFOUT_INT} is not in HH format."
+  exit 1
+elif [ ! ${WRFOUT_INT} -gt 0 ]; then
+  echo "ERROR: \${WRFOUT_INT} must be an integer for the max WRF domain index > 0." 
+  exit 1
+fi
+
+if [ ${#CYCLE_INT} -ne 2 ]; then
+  echo "ERROR: \${CYCLE_INT}, ${CYCLE_INT}, is not in 'HH' format."
+  exit 1
+fi
+
+if [[ ${WRF_IC} = ${REALEXE} ]]; then
+  echo "WRF initial and boundary conditions sourced from real.exe."
+elif [[ ${WRF_IC} = ${CYCLING} ]]; then
+  msg="WRF initial conditions and boundary conditions sourced from GSI / WRFDA "
+  msg+=" analysis."
+  echo ${msg}
+elif [[ ${WRF_IC} = ${RESTART} ]]; then
+  echo "WRF initial conditions sourced from restart files."
+else
+  msg="ERROR: \${WRF_IC}, ${WRF_IC}, must equal REALEXE, CYCLING or RESTART "
+  msg+=" (case insensitive)."
+  echo ${msg}
   exit 1
 fi
 
@@ -266,7 +253,7 @@ fi
 # WRF_ROOT   = Root directory of a clean WRF build WRF/run directory
 # EXP_CONFIG = Root directory containing sub-directories for namelists
 #              vtables, geogrid data, GSI fix files, etc.
-# CYCLE_HOME = Start time named directory for cycling data containing
+# CYCLE_HME = Start time named directory for cycling data containing
 #              bkg, wpsprd, realprd, wrfprd, wrfdaprd, gsiprd, enkfprd
 # DATA_ROOT  = Directory for all forcing data files, including grib files,
 #              obs files, etc.
@@ -309,13 +296,8 @@ elif [ ! -d ${DATA_ROOT} ]; then
   exit 1
 fi
 
-if [ ! "${MPIRUN}" ]; then
+if [ ! ${MPIRUN} ]; then
   echo "ERROR: \${MPIRUN} is not defined."
-  exit 1
-fi
-
-if [ ! "${N_PROC}" ]; then
-  echo "ERROR: \${N_PROC} is not defined."
   exit 1
 fi
 
@@ -324,7 +306,7 @@ if [ ! ${N_PROC} ]; then
   exit 1
 elif [ ! ${N_PROC} -gt 0 ]; then
   msg="ERROR: The variable \${N_PROC} must be set to the number"
-  msg+=" of processors to run real.exe."
+  msg+=" of processors to run wrf.exe."
   echo ${msg}
   exit 1
 fi
@@ -342,7 +324,7 @@ fi
 #
 ##################################################################################
 
-work_root=${CYCLE_HOME}/wrfprd/ens_${memid}
+work_root=${CYCLE_HME}/wrfprd/ens_${memid}
 mkdir -p ${work_root}
 cmd="cd ${work_root}"
 echo ${cmd}; eval ${cmd}
@@ -361,20 +343,23 @@ for file in ${wrf_dat_files[@]}; do
   echo ${cmd}; eval ${cmd}
 done
 
-# Remove any old WRF outputs in the directory
-cmd="rm -f wrfout_*"
-echo ${cmd}; eval ${cmd}
+if [[ ${WRF_IC} = ${REALEXE} || ${WRF_IC} = ${CYCLING} ]]
+  # Remove any old WRF outputs in the directory from failed runs
+  cmd="rm -f wrfout_*"
+  echo ${cmd}; eval ${cmd}
+  cmd="rm -f wrfrst_*"
+  echo ${cmd}; eval ${cmd}
+fi
 
-# Link WRF initial conditions from WPS real or GSI analysis depending on IF_CYCLING
-dmn=1
+# Link WRF initial conditions
 for dmn in {01..${MAX_DOM}}; do
   wrfinput=wrfinput_d${dmn}
   datestr=`date +%Y-%m-%d_%H_%M_%S -d "${strt_time}"`
   # if cycling AND analyzing this domain, get initial conditions from last analysis
-  if [[ ${IF_CYCLING} = ${YES} && ${dmn} -lt ${DOWN_DOM} ]]; then
+  if [[ ${WRF_IC} = ${CYCLING} && ${dmn} -lt ${DOWN_DOM} ]]; then
     if [[ ${dmn} = 01 ]]; then
       # obtain the boundary files from the lateral boundary update by WRFDA 
-      wrfanlroot=${CYCLE_HOME}/wrfdaprd/lateral_bdy_update/ens_${memid}
+      wrfanlroot=${CYCLE_HME}/wrfdaprd/lateral_bdy_update/ens_${memid}
       wrfbdy=${wrfanlroot}/wrfbdy_d01
       cmd="ln -sf ${wrfbdy} ./wrfbdy_d01"
       echo ${cmd}; eval ${cmd}
@@ -387,10 +372,10 @@ for dmn in {01..${MAX_DOM}}; do
       # Nested domains have boundary conditions defined by parent
       if [ ${memid} -eq 00 ]; then
         # control solution is indexed 00, analyzed with GSI
-        wrfanl_root=${CYCLE_HOME}/gsiprd/d${dmn}
+        wrfanl_root=${CYCLE_HME}/gsiprd/d${dmn}
       else
         # ensemble perturbations are updated with EnKF step
-        wrfanl_root=${CYCLE_HOME}/enkfprd/d${dmn}
+        wrfanl_root=${CYCLE_HME}/enkfprd/d${dmn}
       fi
     fi
 
@@ -398,14 +383,23 @@ for dmn in {01..${MAX_DOM}}; do
     wrfanl=${wrfanlroot}/wrfanl_ens_${memid}_${datestr}
     cmd="ln -sf ${wrfanl} ./${wrfinput}"
     echo ${cmd}; eval ${cmd}
-    if [ ! -r "./${wrfinput}" ]; then
-      echo "ERROR: wrfinput ${wrfanl} does not exist or is not readable."
+
+    if [ ! -r ${wrfinput} ]; then
+      echo "ERROR: wrfinput source ${wrfanl} does not exist or is not readable."
+      exit 1
+    fi
+
+  elif [[ ${WRF_IC} = ${RESTART} ]]
+    # check for restart files at valid start time for each domain
+    wrfrst=${work_root}/wrfrst_d${dmn}_${datestr}
+    if [ ! -r ${wrfrst} ]; then
+      echo "ERROR: wrfrst source ${wrfrst} does not exist or is not readable."
       exit 1
     fi
 
   else
-    # else get initial and boundary conditions from real
-    realroot=${CYCLE_HOME}/realprd/ens_${memid}
+    # else get initial and boundary conditions from real for downscaled domains
+    realroot=${CYCLE_HME}/realprd/ens_${memid}
     if [ ${dmn} = 01 ]; then
       # Link the wrfbdy_d01 file from real
       wrfbdy=${realroot}/wrfbdy_d01
@@ -430,7 +424,7 @@ for dmn in {01..${MAX_DOM}}; do
   # NOTE: THIS LINKS SST UPDATE FILES FROM REAL OUTPUTS REGARDLESS OF GSI CYCLING
   if [[ ${IF_SST_UPDTE} = ${YES} ]]; then
     wrflowinp=wrflowinp_d${dmn}
-    realname=${CYCLE_HOME}/realprd/ens_${memid}/${wrflowinp}
+    realname=${CYCLE_HME}/realprd/ens_${memid}/${wrflowinp}
     cmd="ln -sf ${realname} ./"
     echo ${cmd}; eval ${cmd}
     if [ ! -r ${wrflowinp} ]; then
@@ -442,7 +436,7 @@ done
 
 # Move existing rsl files to a subdir if there are any
 echo "Checking for pre-existing rsl files."
-if [ -f "rsl.out.0000" ]; then
+if [ -f rsl.out.0000 ]; then
   rsldir=rsl.wrf.`ls -l --time-style=+%Y-%m-%d_%H_%M%_S rsl.out.0000 | cut -d" " -f 6`
   mkdir ${rsldir}
   echo "Moving pre-existing rsl files to ${rsldir}."
@@ -514,6 +508,19 @@ cat namelist.input \
   | sed "s/\(HISTORY_INTERVAL\)${EQUAL}HISTORY_INTERVAL/\1 = ${hist_int}, /" \
    > namelist.input.tmp
 mv namelist.input.tmp namelist.input
+
+# Update the restart setting in wrf namelist depending on switch
+if [[ ${WRF_IC} = ${RESTART} ]]; then
+  cat namelist.input \
+    | sed "s/\(RESTART\)${EQUAL}RESTART/\1 = .true./" \
+    > namelist.input.tmp
+  mv namelist.input.tmp namelist.input
+else
+  cat namelist.input \
+    | sed "s/\(RESTART\)${EQUAL}RESTART/\1 = .false./" \
+    > namelist.input.tmp
+  mv namelist.input.tmp namelist.input
+fi
 
 # Update the restart interval in wrf namelist to the end of the fcst_len
 cat namelist.input \
@@ -587,25 +594,20 @@ mv namelist.input.tmp namelist.input
 echo
 echo "EXP_CONFIG   = ${EXP_CONFIG}"
 echo "MEMID        = ${MEMID}"
+echo "CYCLE_HME    = ${CYCLE_HME}"
+echo "STRT TIME    = "`date +"%Y-%m-%d_%H_%M_%S" -d "${strt_time}"`
+echo "END TIME     = "`date +"%Y-%m-%d_%H_%M_%S" -d "${end_time}"`
+echo "WRFOUT_INT   = ${WRFOUT_INT}"
 echo "BKG_DATA     = ${BKG_DATA}"
-echo "CYCLE_HOME   = ${CYCLE_HOME}"
-echo "DATA_ROOT    = ${DATA_ROOT}"
-echo
-echo "FCST_LENGTH  = ${fcst_len}"
-echo "FCST INT     = ${WRFOUT_INT}"
 echo "MAX_DOM      = ${MAX_DOM}"
-echo "IF_CYCLING   = ${IF_CYCLING}"
+echo "WRF_IC       = ${WRF_IC}"
 echo "IF_SST_UPDTE = ${IF_SST_UPDTE}"
 echo "IF_FEEDBACK  = ${IF_FEEDBACK}"
 echo
-echo "STRT TIME    = "`date +"%Y/%m/%d %H_%M_%S" -d "${strt_time}"`
-echo "END TIME     = "`date +"%Y/%m/%d %H_%M_%S" -d "${end_time}"`
-echo "CYCLE TIME   = "`date +"%Y/%m/%d %H_%M_%S" -d "${cycle_time}"`
-echo
-now=`date +%Y%m%d%H%M%S`
+now=`date +%Y-%m-%d_%H_%M_%S`
 echo "wrf started at ${now}."
-
-${MPIRUN} -n ${N_PROC} ${wrf_exe}
+cmd="${MPIRUN} -n ${N_PROC} ${wrf_exe}"
+echo ${cmd}; eval ${cmd}
 
 ##################################################################################
 # Run time error check
@@ -615,9 +617,12 @@ error=$?
 # Save a copy of the RSL files
 rsldir=rsl.wrf.${now}
 mkdir ${rsldir}
-mv rsl.out.* ${rsldir}
-mv rsl.error.* ${rsldir}
-cp namelist.* ${rsldir}
+cmd="mv rsl.out.* ${rsldir}"
+echo ${cmd}; eval ${cmd}
+cmd="mv rsl.error.* ${rsldir}"
+echo ${cmd}; eval ${cmd}
+cmd="mv namelist.* ${rsldir}"
+echo ${cmd}; eval ${cmd}
 
 if [ ${error} -ne 0 ]; then
   echo "ERROR: ${wrf_exe} exited with status ${error}."
@@ -626,7 +631,7 @@ fi
 
 # Look for successful completion messages adjusted for quilting processes
 nsuccess=`cat ${rsldir}/rsl.* | awk '/SUCCESS COMPLETE WRF/' | wc -l`
-(( ntotal=(N_PROC - NIO_GROUPS * NIO_TPG ) * 2 ))
+ntotal=$(( (N_PROC - NIO_GROUPS * NIO_TPG ) * 2 ))
 echo "Found ${nsuccess} of ${ntotal} completion messages"
 if [ ${nsuccess} -ne ${ntotal} ]; then
   msg="ERROR: ${wrf_exe} did not complete successfully, missing completion "
@@ -634,18 +639,16 @@ if [ ${nsuccess} -ne ${ntotal} ]; then
   echo 
 fi
 
-# ensure that the cycle_io/date/bkg directory exists for starting next cycle
-cycle_intv=`date +%H -d "${CYCLE_INT}"`
-datestr=`date +%Y%m%d%H -d "${cycle_time} ${cycle_intv} hours"`
+# ensure that the bkg directory exists in next ${CYCLE_HME}
+datestr=`date +%Y%m%d%H -d "${strt_time} ${CYCLE_INT} hours"`
 new_bkg=${datestr}/bkg/ens_${memid}
-mkdir -p ${CYCLE_HOME}/../${new_bkg}
+cmd="mkdir -p ${CYCLE_HME}/../${new_bkg}"
+echo ${cmd}; eval ${cmd}
 
 # Check for all wrfout files on WRFOUT_INT and link files to
 # the appropriate bkg directory
-dmn=1
-while [ ${dmn} -le ${MAX_DOM} ]; do
-  fcst=0
-  while [ ${fcst} -le ${fcst_len} ]; do
+for dmn in {01..${MAX_DOM}}; do
+  for fcst in {000..${fcst_len}..${WRFOUT_INT}}; do
     datestr=`date +%Y-%m-%d_%H_%M_%S -d "${strt_time} ${fcst} hours"`
     if [ ! -s "wrfout_d${dmn}_${datestr}" ]; then
       msg="WRF failed to complete, wrfout_d${dmn}_${datestr} "
@@ -653,10 +656,9 @@ while [ ${dmn} -le ${MAX_DOM} ]; do
       echo ${msg}
       exit 1
     else
-      ln -sfr wrfout_d${dmn}_${datestr} ${CYCLE_HOME}/../${new_bkg}/
+      cmd="ln -sf wrfout_d${dmn}_${datestr} ${CYCLE_HME}/../${new_bkg}"
+      echo ${cmd}; eval ${cmd}
     fi
-
-    (( fcst += WRFOUT_INT ))
   done
 
   if [ ! -s "wrfrst_d${dmn}_${datestr}" ]; then
@@ -665,15 +667,15 @@ while [ ${dmn} -le ${MAX_DOM} ]; do
     echo 
     exit 1
   else
-    ln -sfr wrfrst_d${dmn}_${datestr} ${CYCLE_HOME}/../${new_bkg}/
+    cmd="ln -sf wrfrst_d${dmn}_${datestr} ${CYCLE_HME}/../${new_bkg}"
+    echo ${cmd}; eval ${cmd}
   fi
-
-  (( dmn += 1 ))
 done
 
 # Remove links to the WRF DAT files
 for file in ${wrf_dat_files[@]}; do
-    rm -f `basename ${file}`
+    cmd="rm -f `basename ${file}`"
+    echo ${cmd}; eval ${cmd}
 done
 
 echo "wrf.sh completed successfully at `date`."
